# reuse regex matcher in dictionary based LIKE queries